+++ src/billboard_combat.c
@@ -4,6 +4,9 @@
 #define ENEMY_RADIUS 24
 
 static bool is_position_blocked(s32 x, s32 y) {
+#if DEBUG_PERF
+    bsp_debug_set_query_owner(BSP_QUERY_ENEMY);
+#endif
     return bsp_circle_blocked(x, y, ENEMY_RADIUS);
 }